Implemented HitBox.cpp. Also, added a debug flag to objects which can be set to enable hitbox drawing.

+++ source/headers/HitBox.hpp
@@ -7,10 +7,10 @@ class HitBox{
 	    int offsetY;
 	    int width;
 	    int height;
-	    float ttl;
+	    double ttl;
 
-	    HitBox(int offsetX, int offsetY, int width, int height, float ttl=-1);
-	    float decTime(float delta); //(if ttl > 0 && ttl - delta <= 0){ttl = ttl - 1;} return ttl
+	    HitBox(int offsetX, int offsetY, int width, int height, double ttl=-1);
+	    double decTime(double delta); //(if ttl > 0 && ttl - delta <= 0){ttl = ttl - 1;} return ttl
 
 };
 #endif
